+++ automation/src/epilot/models/operations/createflow.py
@@ -11,9 +11,12 @@ from typing import Optional
 @dataclasses.dataclass
 class CreateFlowResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     automation_flow: Optional[shared_automationflow.AutomationFlow] = dataclasses.field(default=None)
     r"""The created automation flow"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""