+++ automation/src/epilot/models/operations/deleteflow.py
@@ -20,9 +20,12 @@ class DeleteFlowRequest:
 @dataclasses.dataclass
 class DeleteFlowResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     automation_flow: Optional[shared_automationflow.AutomationFlow] = dataclasses.field(default=None)
     r"""The deleted automation flow"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""