+++ automation/src/epilot/models/operations/getexecutions.py
@@ -23,9 +23,12 @@ class GetExecutionsRequest:
 @dataclasses.dataclass
 class GetExecutionsResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     get_executions_resp: Optional[shared_getexecutionsresp.GetExecutionsResp] = dataclasses.field(default=None)
     r"""List of automation executions"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""