+++ automation/src/epilot/models/operations/putflow.py
@@ -22,9 +22,12 @@ class PutFlowRequest:
 @dataclasses.dataclass
 class PutFlowResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     automation_flow: Optional[shared_automationflow.AutomationFlow] = dataclasses.field(default=None)
     r"""The updated automation flow"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""