+++ automation/src/epilot/models/operations/searchflows.py
@@ -26,8 +26,11 @@ class SearchFlowsRequest:
 @dataclasses.dataclass
 class SearchFlowsResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     search_automations_resp: Optional[shared_searchautomationsresp.SearchAutomationsResp] = dataclasses.field(default=None)
     r"""List of automation flows, including total count"""